requirements.txt
Appium-Python-Client==1.0.2
atomicwrites==1.4.1
attrs==23.2.0
colorama==0.4.6
iniconfig==2.0.0
packaging==23.2
pluggy==0.13.1
py==1.11.0
pytest==6.1.1
python-dotenv==1.0.0
selenium==3.9.0
toml==0.10.2
urllib3==2.1.0
